move propertyfilter.js into propertyfilter folder and delete extra propertyfilter.js
